gskslexpression: Fold constant expressions into SPIRV
If an expression is constant, always get the constant expression and
write it to the SPIRV file instead of doing an evaluation.

This has immense benefits when dealing with initializations, because
vec4(1, 2, 3, 4) will actually result in a constant instead of first
creating 4 integers, manually converting each of them to float before
calling the constructor. Yuck.